change: 一般プラグイン-カテゴリ設定の共通化
